fix compile

+++ bcos-cpp-sdk/multigroup/GroupInfo.h
@@ -96,7 +96,7 @@ public:
 
     virtual void setGroupID(std::string const& _groupID) { m_groupID = _groupID; }
     virtual void setChainID(std::string const& _chainID) { m_chainID = _chainID; }
-    virtual ssize_t nodesNum() const
+    virtual int64_t nodesNum() const
     {
         boostssl::utilities::ReadGuard l(x_nodeInfos);
         return m_nodeInfos.size();
